# crowdsource-api/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Crowdsource API",
    "description": "Swagger API Documentation for Crowdsource",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/aggregate-data-count/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "byLanguage",
            "in": "query",
            "type": "string"
          },
          {
            "name": "byState",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
...
    "/v2/stats/summary/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/stats/summary/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/stats/categories/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "language",
            "in": "query",
            "type": "string"
          },
          {
            "name": "timeframe",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/stats/contributions/age/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "language",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/stats/contributions/gender/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "language",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
    "/timeline/{type}": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "language",
            "in": "query",
            "type": "string"
          },
          {
            "name": "timeframe",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        },
        "deprecated": true
      }
    },
...
  },
...
}